+++ src/components/SinglePrintButtonForPickSlip/SinglePrintButtonForPickSlip.js
@@ -13,6 +13,7 @@ const SinglePrintButtonForPickSlip = ({
   pickSlipsToCheck,
   pickSlipsPrintTemplate,
   pickSlipsData,
+  isViewPrintDetailsEnabled,
   onBeforeGetContentForSinglePrintButton,
   onBeforePrintForSinglePrintButton,
   getPrintContentRef,
@@ -26,6 +27,7 @@ const SinglePrintButtonForPickSlip = ({
       disabled={disabled}
       contentRef={getPrintContentRef(request.id)}
       requestId={request.id}
+      isViewPrintDetailsEnabled={isViewPrintDetailsEnabled}
       onBeforeGetContent={onBeforeGetContentForSinglePrintButton}
       onBeforePrint={onBeforePrintForSinglePrintButton}
     >
@@ -46,6 +48,7 @@ SinglePrintButtonForPickSlip.propTypes = {
   pickSlipsToCheck: PropTypes.object.isRequired,
   pickSlipsPrintTemplate: PropTypes.object.isRequired,
   pickSlipsData: PropTypes.object.isRequired,
+  isViewPrintDetailsEnabled: PropTypes.bool.isRequired,
   onBeforeGetContentForSinglePrintButton: PropTypes.func.isRequired,
   onBeforePrintForSinglePrintButton: PropTypes.func.isRequired,
   getPrintContentRef: PropTypes.object.isRequired,