+++ src/components/SinglePrintButtonForPickSlip/SinglePrintButtonForPickSlip.test.js
@@ -59,6 +59,7 @@ describe('SinglePrintButtonForPickSlip', () => {
         },
       }
     ],
+    isViewPrintDetailsEnabled: true,
     onBeforeGetContentForSinglePrintButton: jest.fn(),
     onBeforePrintForSinglePrintButton: jest.fn(),
     getPrintContentRef: mockAccordionStatusRef,